requirements.txt
numpy==1.23.5
protobuf==3.20.*
tensorflow==2.8.0
pillow
